# substationDataServer/src/Org.OpenAPITools/wwwroot/openapi-original.json
{
  "openapi" : "3.0.2",
  "info" : {
    "title" : "Public Substation API for the Smart Service Intelligent Autonomous Grid",
    "description" : "Public API for registered users to access the device data, production data on a device, bay or substation base.",
    "contact" : {
      "name" : "Istvan Mudrak",
      "url" : "https://www.smartservice.energy",
      "email" : "istvan.mudrak@smartservice.energy"
    },
    "license" : {
      "name" : "Proprietary",
      "url" : "https://www.smartservice.energy/licenses/Substation"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://api.smartservice.energy/substation/v1"
  } ],
...
  "paths" : {
    "/substations" : {
      "get" : {
        "tags" : [ "substation" ],
        "operationId" : "findSubstationAll",
        "parameters" : [ {
          "name" : "clientId",
          "in" : "query",
          "description" : "Id of client",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "List of all substations owned by a client. ClientId is proided after authorization. All the below GET requests are working on the same philosophy.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Substations"
                }
              }
            }
          },
          "default" : {
            "description" : "Invalid status",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/substations/{substationId}" : {
      "get" : {
        "tags" : [ "substation" ],
        "operationId" : "findSubstationBySubstationId",
        "parameters" : [ {
          "name" : "clientId",
          "in" : "query",
          "description" : "Id of client",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "substationId",
          "in" : "path",
          "description" : "Id of a given substation",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Substation data idetified by {substationId}",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Substation"
                }
              }
            }
          },
          "default" : {
            "description" : "Invalid status",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Error" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "text" : {
            "type" : "string"
          }
        }
      },
      "Location" : {
        "type" : "object",
        "properties" : {
          "latitude" : {
            "type" : "number",
            "format" : "float"
          },
          "longitude" : {
            "type" : "number",
            "format" : "float"
          },
          "altitude" : {
            "type" : "number",
            "format" : "float"
          }
        },
        "example" : {
          "altitude" : 1.4658129,
          "latitude" : 0.8008282,
          "longitude" : 6.0274563
        }
      },
      "Address" : {
        "type" : "object",
        "properties" : {
          "number" : {
            "type" : "string"
          },
          "street" : {
            "type" : "string"
          },
          "city" : {
            "type" : "string"
          },
          "locality" : {
            "type" : "string"
          },
          "zip" : {
            "type" : "string"
          },
          "country" : {
            "type" : "string"
          }
        },
        "example" : {
          "zip" : "zip",
          "number" : "number",
          "country" : "country",
          "city" : "city",
          "street" : "street",
          "locality" : "locality"
        }
      },
      "Contact" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "phone" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          }
        },
        "example" : {
          "phone" : "phone",
          "name" : "name",
          "email" : "email"
        }
      },
      "Substation" : {
        "type" : "object",
        "properties" : {
          "mrid" : {
            "type" : "string"
          },
          "location" : {
            "$ref" : "#/components/schemas/Location"
          },
          "contact" : {
            "$ref" : "#/components/schemas/Contact"
          },
          "address" : {
            "$ref" : "#/components/schemas/Address"
          },
          "description" : {
            "type" : "string"
          }
        },
        "example" : {
          "mrid" : "mrid",
          "address" : {
            "zip" : "zip",
            "number" : "number",
            "country" : "country",
            "city" : "city",
            "street" : "street",
            "locality" : "locality"
          },
          "contact" : {
            "phone" : "phone",
            "name" : "name",
            "email" : "email"
          },
          "description" : "description",
          "location" : {
            "altitude" : 1.4658129,
            "latitude" : 0.8008282,
            "longitude" : 6.0274563
          }
        }
      },
      "Substations" : {
        "type" : "array",
        "items" : {
          "$ref" : "#/components/schemas/Substation"
        }
      },
...
    },
    "securitySchemes" : {
      "smartservice_auth" : {
        "type" : "oauth2",
        "flows" : {
          "authorizationCode" : {
            "authorizationUrl" : "https://login.microsoftonline.com/f3b28f26-6d10-4522-ab26-3141516cbb1a/oauth2/authorize",
            "tokenUrl" : "https://login.microsoftonline.com/f3b28f26-6d10-4522-ab26-3141516cbb1a/oauth2/token"
          }
        }
      }
    }
  }
}